requirements.txt
Flask
psycopg2-binary
python-dotenv   # 로컬 개발용
